Handle unexpected alerts
Change WebDriverBackedEmbeddedBrowser to handle the unexpected alerts
(when obtaining the source and the current URL) and retry the action
when that happens.
Add interface UnexpectedAlertHandler to allow customisation of the
behaviour when handling the unexpected alerts, for example, at the
moment Firefox requires (explicit) acceptance of the alerts while other
browsers don't (Chrome or PhantomJS). Also, add default implementation
which allows to continue with the actions and does not handle the
alerts.
Change WebDriverBrowserBuilder to set up capability to automatically
accept the unexpected alerts and to use a custom UnexpectedAlertHandler
with Firefox (as it does not yet support the new WebDriver capability).
Update tests (PopUpTest and "popup/index.html") to assert the new
behaviour.

Part of zaproxy/zaproxy#3412 - Ajax Spider closes the browser when an
alert box is triggered